pyproject.toml
[tool.poetry]
name = "utify"
version = "0.1.0"
description = ""
authors = ["yajiez <yajiez.me@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
ipython = "^7.10.2"
fire = "^0.2.1"
tqdm = "^4.40.2"
wasabi = "^0.4.2"
numpy = "^1.17.4"
pandas = "^0.25.3"
pyarrow = "^0.15.1"
matplotlib = "^3.1.2"
seaborn = "^0.9.0"
pillow = "^6.2.1"
sqlalchemy = "^1.3.12"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
sphinx = "^2.3.0"
sphinx_rtd_theme = "^0.4.3"
recommonmark = "^0.6.0"
nbsphinx = "^0.5.0"

[tool.poetry.scripts]
utify = "utify.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
